Split up driver into separate test files
Driver is split up to make test files more readable.
